requirements.txt
altgraph==0.17.3
attrs==22.2.0
Babel==2.11.0
certifi==2022.12.7
chardet==5.0.0
charset-normalizer==2.0.0
colorama==0.4.5
contextlib2==21.6.0
et-xmlfile==1.1.0
exceptiongroup==0.0.0a0
execnet==1.9.0
future==0.18.2
idna==3.4
importlib-metadata==4.8.3
iniconfig==1.1.1
lxml==4.9.1
mock==5.0.1
MouseInfo==0.1.3
numpy==1.19.5
openpyxl==3.0.10
packaging==21.3
pandas==1.1.5
path==16.2.0
path.py==12.5.0
pefile==2022.5.30
Pillow==8.4.0
pluggy==1.0.0
py-calendars==0.0.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==7.0.1
pytest-shutil==1.7.0
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.4
pywin32==304
pywin32-ctypes==0.2.0
requests==2.27.0
six==1.16.0
termcolor==1.1.0
tomli==1.2.3
typing_extensions==4.1.1
urllib3==1.26.14
urlopen==1.0.0
xlsx2html==0.4.0
zipp==3.6.0